requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-recaptcha==2.0.6
django-sslserver==0.22
idna==2.10
Pillow==8.0.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.2
